Create awsome_strategy.py
Awsome_strategy.py uses trading indicators such as Super Trend and SMA5 and SMA35 to find the cross-over similar to that done in the Awsome Ossilitator. It generates BUY/SELL signals and updates the Pandas data frame and the PlaceOrder function will pick the right buy and sell from the stock_analysis function. 

The file is still in work-in-progress and not integrated yet with the Pre-Open_market function.